feat: change migration to support internationalization

+++ src/main/java/com/vi/migrationtool/consultingtypeservice/MigrateConsultingTypeDescriptionToTopicMigrationTask.java
@@ -75,7 +75,7 @@ public class MigrateConsultingTypeDescriptionToTopicMigrationTask extends Migrat
                       ct.getUrls() != null ? ct.getUrls().getRegistrationPostcodeFallbackUrl() : "",
                       ct.getSendFurtherStepsMessage(),
                         "{ \"de\": \"" + ct.getTitles().getShort() + "\"}",
-                        "{ \"de\": \"" + ct.getTitles().getShort() + "\"}",
+                        "{ \"de\": \"" + ct.getTitles().getLong() + "\"}",
                         ct.getTitles().getWelcome(),
                       ct.getTitles().getRegistrationDropdown(),
                       ct.getSlug()
